# docs/openapi.yaml
openapi: 3.0.3
info:
  title: dPagès API v1
  version: '1.0'
  description: >
    API interna del sistema de gestión de comandas de dPagès. Fuente única de
    verdad del contrato entre backend y frontend — ver también
    docs/contrato-api.md para las explicaciones de negocio detrás de cada
    regla (por qué el empaquetado exige valores obligatorios, qué es
    kgEditable, etc.).

    Refleja el estado confirmado hasta la capa 11 (detalle de incidencias,
    CORS, alta manual de cliente, dirección de entrega del pedido, fecha y
    observación por línea, código de tarifa) MÁS las decisiones funcionales
    confirmadas con el cliente el 18/08/2026: campos de agrupación de
    artículo (formato, envasado, agrupación de producción), agrupación de
    rendimiento por categoría, Rendiments Porcs, Panell Producció (shape
    cerrado; las fórmulas numéricas todavía no — ver
    docs/contrato-api.md sección 4.10), Panell Obrador por línea individual
    (ya no agrupado por producto), orígenes de pedido como tabla mantenible
    (ya no un enum fijo), y usuarios/roles. Esta versión es SOLO contrato —
    los endpoints nuevos (rendiments-porcs, panells/produccio,
    origens-comanda, usuaris, rols, jo) todavía no están implementados en
    el backend real.
  contact:
    name: VisioFlow Tech
servers:
  - url: http://localhost:8080
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/tarifes/matriu:
    get:
      tags: [Tarifes]
      summary: Matriz de precios (tarifa × artículo)
      parameters:
        - name: categoriaId
          in: query
          schema: { type: integer }
        - name: cerca
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Mida'
      responses:
        '200':
          description: Matriz paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  tarifes:
                    type: array
                    items: { $ref: '#/components/schemas/TarifaResum' }
                  dades:
                    type: array
                    items:
                      type: object
                      properties:
                        producteId: { type: integer }
                        codi: { type: string }
                        descripcio: { type: string }
                        preus:
                          type: object
                          additionalProperties:
                            type: string
                            nullable: true
                          example: { '1': '9.86', '2': null }
                  paginacio: { $ref: '#/components/schemas/Paginacio' }
  /api/v1/tarifes:
    post:
      tags: [Tarifes]
      summary: Crear una tarifa
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TarifaEntrada' }
      responses:
        '201':
          description: Tarifa creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TarifaResum' }
        '400': { $ref: '#/components/responses/Validacio' }
        '409': { $ref: '#/components/responses/Conflicte' }
  /api/v1/tarifes/{tarifaId}/preus/{producteId}:
    patch:
      tags: [Tarifes]
      summary: Actualizar el precio de un artículo en una tarifa (edición de celda)
      parameters:
        - name: tarifaId
          in: path
          required: true
          schema: { type: integer }
        - name: producteId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preu]
              properties:
                preu: { type: string, example: '9.50' }
      responses:
        '200':
          description: Precio actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  tarifaId: { type: integer }
                  producteId: { type: integer }
                  preu: { type: string }
        '404': { $ref: '#/components/responses/NoTrobat' }
...
  /api/v1/transportistes:
    get:
      tags: [Transportistes]
      summary: Listar transportistas
      responses:
        '200':
          description: Listado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dades:
                    type: array
                    items: { $ref: '#/components/schemas/Transportista' }
    post:
      tags: [Transportistes]
      summary: Crear un transportista
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransportistaEntrada' }
      responses:
        '201':
          description: Transportista creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transportista' }
        '400': { $ref: '#/components/responses/Validacio' }
        '409': { $ref: '#/components/responses/Conflicte' }
  /api/v1/transportistes/{id}:
    patch:
      tags: [Transportistes]
      summary: Actualizar un transportista (parcial)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransportistaEntrada' }
      responses:
        '200':
          description: Transportista actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transportista' }
        '400': { $ref: '#/components/responses/Validacio' }
        '404': { $ref: '#/components/responses/NoTrobat' }
        '409': { $ref: '#/components/responses/Conflicte' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: >
        Requerido en todas las rutas excepto /salut. En desarrollo local con
        AUTH_DISABLED=true no hace falta enviarlo.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Pagina:
      name: pagina
      in: query
      schema: { type: integer, default: 1 }
    Mida:
      name: mida
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  responses:
    NoTrobat:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validacio:
      description: Error de validación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflicte:
      description: Conflicto (p. ej. pedido congelado)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            codi:
              type: string
              enum: [VALIDACIO, NO_AUTENTICAT, SENSE_PERMIS, NO_TROBAT, CONFLICTE, ERROR_INTERN]
            missatge: { type: string }
            detalls:
              type: array
              items:
                type: object
                properties:
                  camp: { type: string }
                  missatge: { type: string }
    Paginacio:
      type: object
      properties:
        pagina: { type: integer }
        mida: { type: integer }
        total: { type: integer }
        totalPagines: { type: integer }
...
    TarifaResum:
      type: object
      description: Forma usada en la matriz de tarifas (GET /tarifes/matriu), con código propio.
      properties:
        id: { type: integer }
        codi: { type: string }
        nom: { type: string }
    TarifaEntrada:
      type: object
      required: [codi, nom]
      description: Cuerpo de POST /tarifes. codi es único — repetido responde 409.
      properties:
        codi: { type: string }
        nom: { type: string }
...
    Transportista:
      type: object
      properties:
        id: { type: integer }
        codi:
          type: string
          nullable: true
          description: 'Texto libre, nemotécnico, definido por el usuario al dar de alta — ej. "TR-DHL".'
        nom: { type: string }
        actiu: { type: boolean }
    TransportistaEntrada:
      type: object
      description: >-
        Cuerpo de POST /transportistes (nom obligatorio) y PATCH
        /transportistes/{id} (cuerpo parcial). codi es opcional pero único
        cuando se manda — repetido responde 409.
      properties:
        nom: { type: string }
        codi: { type: string, nullable: true }
...